+++ test/devices_tests/numeric_value_test.py
@@ -256,6 +256,36 @@ class TestNumericValue:
         await sensor.process(telegram)
         after_update_callback.assert_called_with(sensor)
         assert sensor.last_telegram == telegram
+        # consecutive telegrams with same payload shall only trigger one callback
+        after_update_callback.reset_mock()
+        await sensor.process(telegram)
+        after_update_callback.assert_not_called()
+
+    async def test_process_callback_always(self):
+        """Test process / reading telegrams from telegram queue. Test if callback is called."""
+
+        xknx = XKNX()
+        sensor = NumericValue(
+            xknx,
+            "TestSensor",
+            group_address="1/2/3",
+            value_type="temperature",
+            always_callback=True,
+        )
+        after_update_callback = AsyncMock()
+        sensor.register_device_updated_cb(after_update_callback)
+
+        telegram = Telegram(
+            destination_address=GroupAddress("1/2/3"),
+            payload=GroupValueWrite(DPTArray((0x01, 0x02))),
+        )
+        await sensor.process(telegram)
+        after_update_callback.assert_called_with(sensor)
+        assert sensor.last_telegram == telegram
+        # every telegram shall trigger callback
+        after_update_callback.reset_mock()
+        await sensor.process(telegram)
+        after_update_callback.assert_called_with(sensor)
 
     async def test_process_callback_set(self):
         """Test setting value. Test if callback is called."""